🎨 style #31: OrangeButton, InputView 파일분리
