+++ ios/SecondHand/SecondHand/Profile/SignInViewController.swift
@@ -80,105 +80,3 @@ class SignInViewController: UIViewController {
         self.present(signUpViewController, animated: true)
     }
 }
-
-// TODO: 파일 분리 예정 - 1
-class InputView: UIView {
-    
-    let horizontalStackView: UIStackView = {
-        let uiStackView = UIStackView()
-        uiStackView.axis = .horizontal
-        uiStackView.spacing = 8
-        uiStackView.alignment = .fill
-        uiStackView.layoutMargins = UIEdgeInsets(top: 0, left: 8, bottom: 0, right: 8)
-        uiStackView.isLayoutMarginsRelativeArrangement = true
-        
-        
-        uiStackView.translatesAutoresizingMaskIntoConstraints = false
-        return uiStackView
-    }()
-    
-    let textLabel: UILabel = {
-        let label = UILabel()
-        label.translatesAutoresizingMaskIntoConstraints = false
-        label.textAlignment = .left
-        return label
-    }()
-    
-    let inputField: UITextField = {
-        let textField = UITextField()
-        textField.translatesAutoresizingMaskIntoConstraints = false
-        
-        return textField
-    }()
-    
-    override init(frame: CGRect) {
-        super.init(frame: frame)
-        self.setup()
-    }
-    
-    required init?(coder: NSCoder) {
-        super.init(coder: coder)
-        self.setup()
-    }
-    
-    
-    func configureText(labelText: String, textFieldPlaceholder: String) {
-        self.textLabel.text = labelText
-        self.inputField.placeholder = textFieldPlaceholder
-    }
-    
-    private func setup() {
-        self.configureLayout()
-    }
-    
-    private func configureLayout() {
-        self.addSubview(self.horizontalStackView)
-        self.horizontalStackView.addArrangedSubview(textLabel)
-        self.horizontalStackView.addArrangedSubview(inputField)
-
-        NSLayoutConstraint.activate([
-            textLabel.widthAnchor.constraint(equalToConstant: 100),
-            horizontalStackView.topAnchor.constraint(equalTo: self.topAnchor),
-            horizontalStackView.leadingAnchor.constraint(equalTo: self.leadingAnchor),
-            horizontalStackView.trailingAnchor.constraint(equalTo: self.trailingAnchor),
-            horizontalStackView.bottomAnchor.constraint(equalTo: self.bottomAnchor),
-        ])
-    }
-}
-
-// TODO: 파일 분리 예정 - 2
-class OrangeButton: UIButton {
-    
-    override init(frame: CGRect) {
-        super.init(frame: frame)
-        self.setup()
-    }
-    
-    required init?(coder: NSCoder) {
-        super.init(coder: coder)
-        self.setup()
-    }
-    
-    private func setup() {
-        self.setColor()
-        self.configureLayout()
-        self.configureShape()
-    }
-    
-    private func setColor() {
-        self.backgroundColor = UIColor(named: "orange")
-    }
-    
-    private func configureLayout() {
-        self.translatesAutoresizingMaskIntoConstraints = false
-        NSLayoutConstraint.activate([
-            self.widthAnchor.constraint(equalToConstant: 330),
-            self.heightAnchor.constraint(equalToConstant: 40)
-        ])
-    }
-    
-    private func configureShape() {
-        self.layer.cornerRadius = 10
-    }
-    
-}
